Unify CS loading detection; avoid side effects
